+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.2.0
 =====
 
+* Support for assuming the role defined by `role_arn` in the shared config file
 * Support `duration_seconds` option in shared config file to default the token lifetime
 * Export new `AWS_SESSION_EXPIRY` environment variable indicating when the token expires.